Simulated sample surveys drawn from the MU284 Population using a complex design

A list with 10 sample surveys with sample size 15 drawn from the MU284 dataset using a complex sampling design.

Format

A list with 10 data frames, each with 15 rows and 11 columns:

LABEL, ..., CL

Same as MU284 dataset

sample_weight

The sampling weight for the row

Details

The sampling design comes from Ex. 4.3.2 (pg 142-3) of 'Model Assisted Survey Sampling' by Sarndal, Swensson, and Wretman (2003, ISBN:0387406204).

The design is a two-stage sample:

stage I: the primary sampling units (PSUs) are the standard clusters from MU284; we take a simple random sample without replacement of n_I = 5 out of N_I = 50 of these

stage II: within each sampled PSU, we take a simple random sample without replacement of n_i = 3 out of N_i municipalities