Survey bootstrap variance estimators

surveybootstrap has methods for analyzing data that were collected using network reporting techniques. It includes estimators appropriate for the simple bootstrap and the rescaled bootstrap.

See also

Author

Maintainer: Dennis M. Feehan feehan@berkeley.edu

Other contributors:

Matthew J. Salganik mjs3@princeton.edu [thesis advisor]